test-plant/plantguard/requirements.txt
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
Pillow==10.3.0
PyYAML==6.0.1
tqdm==4.66.4
grad-cam==1.5.0
kagglehub==0.3.3
jupyter==1.0.0
